load deja vu sans font from cached file
